Clean up emit for size truncation.  Division here ensures we fit nicely.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@157946 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/ssl/ssl_engine_vars.c
@@ -489,7 +489,7 @@ static char *ssl_var_lookup_ssl_cert_remain(apr_pool_t *p, ASN1_UTCTIME *tm)
         return apr_pstrdup(p, "0");
     }
     
-    diff = (apr_time_sec(then) - apr_time_sec(now)) / (60*60*24);
+    diff = (long)(apr_time_sec(then) - apr_time_sec(now)) / (60*60*24));
 
     return diff > 0 ? apr_ltoa(p, diff) : apr_pstrdup(p, "0");
 }
